Fixed data type conversion

+++ models.go
@@ -19,10 +19,10 @@ type Users [] User
 type Item struct {
 	ID int64			`json:"id"`
 	Title string        `json:"title"`
-	Owner int			`json:"owner"`
+	Owner int64			`json:"owner"`
 	Place string		`json:"place"`
 	Extra string		`json:"extra"`
-	Price float32		`json:"price"`
+	Price float64		`json:"price"`
 	CreatedOn time.Time	`json:"createdOn"`
 }
 // Slice of items
